+++ src/main/java/service/pdf/dto/TocEntry.java
@@ -10,6 +10,6 @@ package service.pdf.dto;
 public record TocEntry(
         int page,
         String chapter,
-        int possiblePoints
+        double possiblePoints
 ) {
 }
